Got the refactor to compile
