+++ SpockOpaleye/app/Main.hs
@@ -16,6 +16,7 @@ import           Web.Spock.Config
 
 import           Control.Monad.Reader
 import           Control.Monad.Writer
+import qualified Control.Monad.Reader as R
 import           CryptoDef
 import qualified Data.Text                  as T
 import           Data.Time
@@ -26,11 +27,27 @@ import           Control.Exception.Lifted
 import           Airbrake
 import           Airbrake.WebRequest
 import           Data.ByteString (ByteString)
+import TenantDefs
+import UserDefs
 
 data MySession =
   EmptySession
 
 data MyAppState = DummyAppState
+
+
+data AppResult a = AppOk a | AppErr T.Text
+
+getCurrentUser :: AppM (Maybe User)
+getCurrentUser = do
+  (_, _, user) <- R.ask
+  return user
+
+
+getCurrentTenant :: AppM (Maybe Tenant)
+getCurrentTenant = do
+  (_, tenant, _) <- R.ask
+  return tenant
 
 connectDb :: IO Connection
 connectDb = connect defaultConnectInfo { connectDatabase = "haskell-webapps" }
