+++ SpockOpaleye/src/ApiBase.hs
@@ -10,6 +10,7 @@ module ApiBase where
 import           Control.Lens
 import           Control.Monad.IO.Class
 import           Control.Monad.Reader
+import qualified Control.Monad.Reader as R
 import           Control.Monad.Writer
 import qualified Data.Profunctor.Product.Default as D
 import           Data.Time                       (UTCTime, getCurrentTime)
@@ -23,13 +24,28 @@ import           TH
 import           Data.Aeson (Value(..))
 import           Data.ByteString (ByteString)
 import           JsonInstances ()
+import DataTypes
+import Lenses
+import           Database.PostgreSQL.Simple
+import UserDefs
+import TenantDefs
 
-makeAuditableLenses ''InternalRole
-makeAuditableLenses ''InternalTenant
-makeAuditableLenses ''InternalUser
+getConnection :: AppM Connection
+getConnection = do
+  (conn, _, _) <- R.ask
+  return conn
 
-auditLog :: ByteString -> AppM ()
-auditLog = tell 
+getCurrentUser :: AppM (Maybe User)
+getCurrentUser = do
+  (_, _, user) <- R.ask
+  return user
+
+
+getCurrentTenant :: AppM (Maybe Tenant)
+getCurrentTenant = do
+  (_, tenant, _) <- R.ask
+  return tenant
+
 
 removeRawDbRows :: Table columnsW columnsR -> (columnsR -> Column PGBool) -> AppM GHC.Int.Int64
 removeRawDbRows table matchFunc = do
