+++ SpockOpaleye/src/JsonInstances.hs
@@ -10,69 +10,14 @@ import           Data.Char
 import           Data.Text
 import           DataTypes
 
+import TH
+import qualified Data.HashMap.Strict as HM
+
+auditable :: a -> Auditable a
+auditable a = Auditable {_data = a, _log = Object HM.empty}
+
+wrapAuditable :: (Functor a, Functor b) => a (b c) -> a (b (Auditable c))
+wrapAuditable a = (fmap auditable) <$> a
+
 instance (FromJSON a) => FromJSON (Auditable a) where
   parseJSON j = auditable <$> (parseJSON j)
-  
-instance FromJSON UserId where
-  parseJSON j@(Number _) = UserId <$> (parseJSON j)
-  parseJSON invalid      = typeMismatch "UserId" invalid
-
-instance FromJSON TenantId where
-  parseJSON j@(Number _) = TenantId <$> (parseJSON j)
-  parseJSON invalid      = typeMismatch "TenantId" invalid
-
-instance FromJSON TenantStatus where
-  parseJSON j@(String _) = tStatus <$> (parseJSON j)
-    where
-      tStatus :: Text -> TenantStatus
-      tStatus "active"   = TenantStatusActive
-      tStatus "inactive" = TenantStatusInActive
-      tStatus "new"      = TenantStatusNew
-      tStatus _      = error "Unknown status name while parsing TenantStatus field"
-  parseJSON invalid = typeMismatch "TenantStatus" invalid
-
-instance FromJSON TenantIncoming where
-  parseJSON (Object v) =
-    (Tenant () () ()) <$> v .: "name" <*> v .: "firstname" <*> v .: "lastname" <*>
-    v .: "email" <*>
-    v .: "phone" <*>
-    (pure ()) <*>
-    v .: "userId" <*>
-    v .: "backofficeDomain"
-  parseJSON invalid = typeMismatch "Unexpected type while paring TenantIncoming" invalid
-
-instance ToJSON TenantStatus where
-  toJSON = genericToJSON defaultOptions
-  toEncoding = genericToEncoding defaultOptions { constructorTagModifier = tgModify }
-    where
-      tgModify :: String -> String
-      tgModify "TenantStatusActive"   = "active"
-      tgModify "TenantStatusInActive" = "inactive"
-      tgModify "TenantStatusNew"      = "new"
-      tgModify _                      = error "Unknown status name for tenant"
-
-instance ToJSON InternalTenant where
-  toJSON = genericToJSON defaultOptions
-  toEncoding = genericToEncoding defaultOptions { fieldLabelModifier = (fmap Data.Char.toLower).removePrefix }
-    where
-      removePrefix = Prelude.drop 11
-
-instance ToJSON UserId where
-  toJSON = genericToJSON defaultOptions
-  toEncoding = genericToEncoding defaultOptions
-
-instance ToJSON TenantId where
-  toJSON = genericToJSON defaultOptions
-  toEncoding = genericToEncoding defaultOptions
-
-instance ToJSON UserStatus where
-  toJSON x = String $ Data.Text.pack $ show x
-
-instance ToJSON RoleId where
-  toJSON (RoleId x) = toJSON x
-
-instance ToJSON Permission where
-  toJSON x = toJSON $ show x
-
-instance (ToJSON a) => ToJSON (Auditable a) where
-  toJSON Auditable {_data = x} = toJSON x
