+++ SpockOpaleye/stack.yaml
@@ -73,3 +73,5 @@ extra-lib-dirs: []
 #
 # Allow a newer minor version of GHC than the snapshot specifies
 # compiler-check: newer-minor
+ghc-options:
+  SpockOpaleye: -ddump-splices
